libxw+setup-env: Add skeleton for libxw + helper script
As of this commit, it should be possible to run

    $ source setup-env
    $ python
    >>> import libxw

without hitting any errors (though `libxw` itself doesn't actually
contain anything interesting yet).

There's a ton of boilerplate here, but basically we just want to
create a virtual environment specific for `crosswords` development
and also follow some Python packaging "best practices".

The main idea with `setup-env` is that we only need to run it once
when we start working (or after adding/updating dependencies).  After
that, changes to source code should be visible inside the virtual
environment automatically.